+++ fire_atlas_editor/CHANGELOG.md
@@ -2,3 +2,4 @@
 
  - Fix Preview widget update
  - Added tooltips to icon buttons
+ - Adding Bitmap automap feature